Struct aurelius::Server

pub struct Server {
    // some fields omitted
}

The Server type constructs a new markdown preview server.

The server will listen for HTTP and WebSocket connections on arbitrary ports.

Methods

impl Server

fn new() -> Server

Creates a new markdown preview server.

Builder methods are provided to configure the server before starting it.

fn initial_markdown(&mut self, markdown: &str) -> &mut Server

Sets the markdown that the server should display when the first connection is received.

fn highlight_theme(&mut self, theme: &str) -> &mut Server

Sets the theme that should be used for syntax highlighting.

Syntax highlighting is provided by highlight.js. All themes supported by highlight.js are supported.

fn start(self) -> ServerHandle

Starts the server, returning a ServerHandle to communicate with it.